+++ src/main/java/racingcar/domain/RacingCars.java
@@ -1,6 +1,6 @@
 package racingcar.domain;
 
-import racingcar.RandomGenerator;
+import racingcar.util.RandomGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
